Improve testing of metadata service
