+++ src/Shark.Fido2.Metadata.Core/Abstractions/Repositories/IHttpClientConformanceTestRepository.cs
@@ -4,7 +4,7 @@ namespace Shark.Fido2.Metadata.Core.Abstractions.Repositories;
 
 public interface IHttpClientConformanceTestRepository
 {
-    Task<List<string>> GetMetadataBlobEndpoints(CancellationToken cancellationToken);
+    Task<List<string>> GetMetadataBlobEndpoints(string remoteUrl, CancellationToken cancellationToken);
     Task<string> GetMetadataBlob(string endpoint, CancellationToken cancellationToken);
-    Task<X509Certificate2> GetRootCertificate(CancellationToken cancellationToken);
+    Task<X509Certificate2> GetRootCertificate(string url, CancellationToken cancellationToken);
 }
