+++ src/Shark.Sample.Fido2/appsettings.Test.json
@@ -16,7 +16,7 @@
     "EnableMetadataService": true,
     "EnableStrictAuthenticatorVerification": true,
     "MetadataServiceConfiguration": {
-      "MetadataBlobLocation": "https://mds3.fido.tools/getEndpoints",
+      "MetadataBlobLocation": "https://mds3.fido.tools/getEndpoints;C:\\src\\fido2\\tests\\Metadata",
       "RootCertificateLocationUrl": "https://mds3.fido.tools/pki/MDS3ROOT.crt",
       "MaximumTokenSizeInBytes": 6291456
     }
