--- src/Shark.Sample.Fido2/appsettings.TestLocalMetadataBlob.json
@@ -1,24 +1,0 @@
-{
-  "DetailedErrors": true,
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "Fido2Configuration": {
-    "Origin": "localhost",
-    "RelyingPartyId": "localhost",
-    "RelyingPartyIdName": "Shark Corporation",
-    "Timeout": "30000",
-    "EnableTrustedExecutionEnvironmentOnly": true,
-    "EnableMetadataService": true,
-    "EnableStrictAuthenticatorVerification": false,
-    "MetadataServiceConfiguration": {
-      "MetadataBlobLocation": "C:\\src\\fido2\\tests\\Metadata",
-      "RootCertificateLocationUrl": null,
-      "MaximumTokenSizeInBytes": 6291456
-    }
-  }
-}
